fix: two bugs in B-tree key lookup
Bug 1 — BranchPage.get() used strict less-than (<) when selecting
which child page to descend into. In LMDB, a branch node's key is the
*first* key of its child page, so the correct condition is ≤ (<=):

  lastOrNull { branchKey.compareWith(searchKey) <= 0 }

With the old < 0 condition, searching for a key that exactly equals
a branch separator key would navigate to the preceding (wrong) child
page, resulting in KeyNotFoundInPage on the leaf.

Bug 2 — compareWith() compared bytes as signed Kotlin Bytes instead
of the unsigned lexicographic order LMDB uses by default. Bytes > 0x7F
sorted incorrectly, breaking get() on any database whose branch keys
contained high bytes (e.g. binary keys). Fix: use toUByte() before
comparison.

Both bugs are covered by the new RandomDatabaseRoundtripTests suite,
which generates real LMDB databases via lmdbjava and verifies that
lmdb-kt can read every key back via both dump() and get().

+++ src/main/kotlin/com/growse/lmdb_kt/Extensions.kt
@@ -22,7 +22,7 @@ internal fun ByteArray.compareWith(byteArray: ByteArray): Int {
   if (byteArray.isEmpty()) return 1
   val firstDifference = this.zip(byteArray).firstOrNull { it.first != it.second }
   return when {
-    firstDifference != null -> if (firstDifference.first > firstDifference.second) 1 else -1
+    firstDifference != null -> if (firstDifference.first.toUByte() > firstDifference.second.toUByte()) 1 else -1
     this.size > byteArray.size -> 1
     else -> -1
   }
